Use .Heartbeat instead of .Stepped
Stepped is arbitrary for use in a scheduler, physic simulation has no relation here, a laggy simulation will not affect this module, eliminating any reason you could come up with to use .Stepped with this.  Switched to .Heartbeat, heartbeat is where things like this are expected to run and also provides more sensible and predictable timing on when scheduled functions could run.